Create Awareness
In our spiritual development, an important aspect involves creating awareness of thoughts, beliefs, attitudes, awareness, actions, and mindsets. Forming beneficial daily rituals will expand our awareness in these areas, and will allow us to become an objective witness to events occurring in our lives. For example, try starting a daily meditation ritual around the same time each day. This practice does not need to be extremely lengthy and can even last 10 minutes. Studies have shown that maintaining a morning gratitude meditation in nature when the sun rises will not only promote peace and tranquility, but will also strengthen awareness to nature and its surroundings, our internal guidance system, and our appreciation for small beauties. Additionally, it is a good idea to track any thoughts, observations, questions, or ideas in a journal so you can go back and review your progress, as well as note any beneficial changes that occur in your life after practicing these rituals for some time. This is a slow and gradual process so please do not get disappointed or try to rush things. Remember, there is no rush to our spiritual development and that we are all exactly where we are meant to be at any given moment. It is up to us to decide how we want to live each journey in life, how we want to grow, and what we are willing to do to get there.